SMBIOS building Fixed
Had to rebase current commit of vl.c to sync up....

That was painful, but that happens when you're patching an old codebase that was slapped together in a 'just make it work 'manner with parts from various versions/times

I don't have the help/options generation yet sorted, so the smbios help is.... a bit... rough. Mainly, it's just not there.