upstream 54586bd
Unfortunately, Visual Studio 17.9 is now the minimum requirement to build, as they implemented what was at the time a GCC extension typeof() which is only now part of the C23 standard and supported in MSVC as of February 2024. Unable to set stdclatest in GUI, we use the __typeof__ macro that allows it to work regardless of standard version.

Also switched compiler to C18 support instead of legacy MSVC support.